refactor: HourlyWeather Extension  분리
